+++ src/src/utils.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int32_t map(int32_t x, int32_t in_min, int32_t in_max, int32_t out_min, int32_t out_max);
+float map(float x, float in_min, float in_max, float out_min, float out_max);
 int32_t constrain(int32_t x, int32_t min, int32_t max);
 
 #endif
